worked on EVAL_APTw_3T.m
The current approach uses soft coded case checks: if you use pulseq seq file, load if from pulseq library. If not, give the seq file to the GUI and check if there is an M_z simulation txt file. If this is also not there, simulate the seq-file again.

Also, simulation and data is plotted

Current FLAW is still, that the mean Z-spectrum is going to be caclulated over the whole phantom/unsegmented head.